Moves yaml constructors back into parse
This doesn't look as tidy, but it helps ensure yaml.safe_load's state is the
same in the YamlConstructorTests as it is in parse.py, where the code is
actually running. Previously, we were adding yaml constructors within
the test_yaml_constructors test suite and in parse.py, which allowed
tests to pass while code failed.